make `create` more like other mirage `connect` functions
MirageOS `connect` functions now throw exceptions on failure conditions
instead of returning an `Error polyvar, as the expected behavior of a
unikernel is to refuse to boot in the presence of such errors and
therefore a programmatic handling isn't expected.  Make `create` fit in
a bit better, and simplify the bootvar code in the process.