convert to topkg; call self mirage-bootvar-xen, not mirage-bootvar
Remove OASIS and replace it with topkg.

`mirage-bootvar-xen` installed a library called `mirage-bootvar`.  So
does `mirage-bootvar-solo5`.  This does not currently cause a problem as
`mirage-bootvar-solo5` and `mirage-bootvar-xen` are not coinstallable in
the same switch due to another problem, but we hope to fix that soon.
Once that happens, `mirage-bootvar-xen` and `mirage-bootvar-solo5` will
need to coexist.