Realign interface with mirage-bootvar-solo5
The additional processing done to support the existing Xen-specific
Bootvar interfaces seems to be interfering with the new PVH-only Xen
stack; command line arguments passed via xl.cfg cmdline= are getting
dropped on the floor.

Since the Xen-specific interfaces don't appear to be used anywhere, and
mirage/mirage can be trivially updated to use only Bootvar.argv()
without filtering, this change drops the Xen-specific interfaces
altogether, effectively aligning the interface with that provided by
mirage-bootvar-solo5, with the sole difference being dependencies on the
Xen stack, and the naming of the external call used to get the raw
command line.